Add demo app
Run `npm start` to start everything up and then open this in the browser:

http://localhost:8086/product-a

That may trigger the reload loop.

It's not 100% reproducible, so if it doesn't reload, try restarting the server.
If the browser page is still trying to ping the Websocket server, it usually
kicks back in and starts the reload loop when the server starts back up.